# Improve issymbollike performance