https://github.com/SciML/ModelingToolkit.jl/issues/1102#issuecomment-876130158


@baggepinnen  This is great!

Old:
```
BechmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  5.215 μs … 467.285 μs  ┊ GC (min … max): 0.00% … 98.36%
 Time  (median):     5.382 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.521 μs ±   6.023 μs  ┊ GC (mean ± σ):  1.53% ±  1.39%

  ▁▃▅██▇▅▄▂▁                                                  ▂
  █████████████▇██▇▇▇▆▄▆▃▆▅▃▃▄▁▅▄▄▃▄▁▃▃▁▅▁▃▃▃▁▃▄▃▄▁▄▄▅▄▁▄▃▃▄▃ █
  5.22 μs      Histogram: log(frequency) by time      7.47 μs <

 Memory estimate: 1.53 KiB, allocs estimate: 35.
```

New:
```
BechmarkTools.Trial: 10000 samples with 109 evaluations.
 Range (min … max):  769.495 ns …  20.644 μs  ┊ GC (min … max): 0.00% … 95.35%
 Time  (median):     794.725 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   805.460 ns ± 382.583 ns  ┊ GC (mean ± σ):  0.94% ±  1.90%

        ▂▄▅██▁                                                   
  ▂▃▃▄▄▅██████▄▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  769 ns           Histogram: frequency by time          939 ns <

 Memory estimate: 224 bytes, allocs estimate: 5.% 
```